Refactor CameraControl
* Separate Definition File Parsing to dedicated class
* Separate HTTP Downloading to dedicated class
* Move Video Stream Info class to own files
* Move Camera Options to own files
